requirements.txt
Flask
requests
python-dotenv
psycopg2-binary
Flask-SQLAlchemy==2.5.1
flask-login
Flask-Migrate
spotipy
flask_sqlalchemy
flask_login
sqlalchemy
flask
